requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==3.0.2
django-timezone-field==4.0
django-treebeard==4.3.1
kombu==3.0.37
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==2.10.6
six==1.13.0
sqlparse==0.3.0
